Move UserProvider into its own file
